requirements.txt
amqp==5.1.1
backports.zoneinfo==0.2.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dotenv
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
mysqlclient==2.1.1
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.15
typing-extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0